refactor(validators): Moved linkValidator from helper function to shared util
