+++ src/app/core/helpers/index.ts
@@ -1,4 +1,3 @@
 export * from './http.helper';
 export * from './i18n.helper';
-export * from './link-validator.helper';
 export * from './types.helper';
